pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
dependency-injector = "^4.41.0"
pymysql = "^1.1.0"
python-dotenv = "^1.0.0"
cryptography = "^45.0.5"
httpx = "^0.25.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
google-generativeai = "^0.8.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

